% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAR_with_constant_threshold.R
\name{TAR_const}
\alias{TAR_const}
\title{Estimate a threshold autoregressive model with constant thresholds.}
\usage{
TAR_const(y)
}
\arguments{
\item{y}{time series vector}
}
\description{
Estimate a threshold autoregressive model with constant thresholds.
}
